Freq of Rank of Non-sustained Complained Policemen
Sixty-six percent of the complaints were non-sustained. Non-sustained cases arise when the allegations are dismissed due to lack of evidence  to prove or disprove the allegations. 
However, it is good to note that none of the ranks captain, chief, commander, deputy chief, chief were disciplined. 
There were a total 29 complaints that involved these said ranks that were not-sustained or in short dismissed. 
This brings up the case that most of the complaints before were investigated by internal staff and were not disclosed to the public which makes the decision fishy.